requirements.txt
opencv-python
albumentations
segmentation-models-pytorch
tqdm
scikit-learn
matplotlib
pandas
torchmetrics
torchinfo
pyyaml